Move UserspaceMemoryRegion struct and related functions into kvm/memory.go
